+++ src/Bugfree.Spo.ExternalSharingCenter.Core/Domain.cs
@@ -102,6 +102,5 @@ namespace Bugfree.Spo.ExternalSharingCenter
         public DateTime Start { get; set; }
         public DateTime End { get; set; }
         public string Comment { get; set; }
-        public DateTime Created { get; set; }
     }
 }