+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Bugfree.Spo.ExternalSharingCenter 1.0.2 (2016-08-04)
 
 * [#4](https://github.com/ronnieholm/Bugfree.Spo.ExternalSharingCenter/issues/4) - Ensure internal consistency between SharePoint lists on External Sharing Center
+* [#5](https://github.com/ronnieholm/Bugfree.Spo.ExternalSharingCenter/issues/5) - Backup internal lists for restore on same or other External Sharing Center web
 
 ## Bugfree.Spo.ExternalSharingCenter 1.0.1 (2016-07-31)